Fix warning: implicit conversion changes signedness
visp/modules/io/src/image/private/vpImageIoPortable.cpp: In function ‘void vp_readPFM_HDR(vpImage<float>&, const std::string&)’:
visp/modules/io/src/image/private/vpImageIoPortable.cpp:517:45: warning: conversion to ‘std::streamsize’ {aka ‘long int’} from ‘long unsigned int’ may change the sign of the result [-Wsign-conversion]
  517 |     fd.read((char *)I[i], sizeof(float) * w * channels);
      |                           ~~~~~~~~~~~~~~~~~~^~~~~~~~~~
visp/modules/io/src/image/private/vpImageIoPortable.cpp: In function ‘void vp_readPFM_HDR(vpImage<vpRGBf>&, const std::string&)’:
visp/modules/io/src/image/private/vpImageIoPortable.cpp:592:45: warning: conversion to ‘std::streamsize’ {aka ‘long int’} from ‘long unsigned int’ may change the sign of the result [-Wsign-conversion]
  592 |     fd.read((char *)I[i], sizeof(float) * w * channels);
      |                           ~~~~~~~~~~~~~~~~~~^~~~~~~~~~

+++ modules/io/src/image/private/vpImageIoPortable.cpp
@@ -514,7 +514,7 @@ void vp_readPFM_HDR(vpImage<float> &I, const std::string &filename)
   bool swapEndianness = littleEndian;
 #endif
   for (int i = static_cast<int>(I.getHeight()) - 1; i >= 0; i--) {
-    fd.read((char *)I[i], sizeof(float) * w * channels);
+    fd.read((char *)I[i], static_cast<unsigned int>(sizeof(float) * w * channels));
     if (swapEndianness) {
       for (unsigned int j = 0; j < w * channels; ++j) {
         I[i][j] = vpEndian::swapFloat(I[i][j]);
@@ -589,7 +589,7 @@ void vp_readPFM_HDR(vpImage<vpRGBf> &I, const std::string &filename)
   bool swapEndianness = littleEndian;
 #endif
   for (int i = static_cast<int>(I.getHeight()) - 1; i >= 0; i--) {
-    fd.read((char *)I[i], sizeof(float) * w * channels);
+    fd.read((char *)I[i], static_cast<unsigned int>(sizeof(float) * w * channels));
     if (swapEndianness) {
       for (unsigned int j = 0; j < w; ++j) {
         I[i][j].R = vpEndian::swapFloat(I[i][j].R);
